chore(confettis): refactoring confetti components
