+++ src/components/index.ts
@@ -11,3 +11,4 @@ export * from './EmojiPopper';
 export * from './Emoji';
 export * from './Popper';
 export * from './Confettis';
+export * from './Confetti';
