+++ src/utils/confettis.ts
@@ -1,0 +1,9 @@
+export function degreesToRadians(degrees: number) {
+  var pi = Math.PI;
+  return degrees * (pi / 180);
+}
+
+export function randomIntFromInterval(min: number, max: number): number {
+  // min and max included
+  return Math.floor(Math.random() * (max - min + 1) + min);
+}
